@@ -6,11 +6,9 @@
   "browser": "./dist/browser/logupts.min.js",
   "types": "./dist/es2015/logupts.d.ts",
   "scripts": {
-    "test": "rm -Rf log && mocha -r node_modules/ts-node/register src/**/*.spec.ts",
-    "watch": "tsc --watch & rollup -c --watch",
-    "karma": "karma start",
-    "build": "sh ./shellscripts/build.sh",
-    "externals": "sh ./shellscripts/externals.sh"
+    "build": "tsc && rollup -c",
+    "test": "TS_NODE_COMPILER_OPTIONS='{\"module\":\"commonjs\"}' mocha --opts ./mocha.opts",
+    "coverage": "tsc && nyc npm test"
   },
   "keywords": [
     "logging",
@@ -30,24 +28,22 @@
   },
   "homepage": "https://github.com/milleniumfrog/logupts#readme",
   "devDependencies": {
+    "@istanbuljs/nyc-config-typescript": "^0.1.3",
     "@types/chai": "^4.1.7",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.10",
+    "@types/mocha": "^5.2.6",
+    "@types/node": "^12.0.0",
     "chai": "^4.2.0",
-    "karma": "^2.0.5",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-mocha": "^1.3.0",
-    "karma-typescript": "^3.0.13",
-    "mapcat": "^0.4.0",
-    "mocha": "^5.2.0",
-    "rollup": "^0.63.5",
-    "rollup-plugin-commonjs": "^9.2.0",
-    "rollup-plugin-node-resolve": "^3.4.0",
-    "strplace": "0.0.4",
-    "ts-node": "^7.0.1",
-    "typedoc": "^0.12.0",
-    "typescript": "^2.9.2",
-    "uglify-es": "^3.3.9"
+    "mocha": "^6.1.4",
+    "nyc": "^14.1.0",
+    "rollup": "^1.11.3",
+    "rollup-plugin-node-resolve": "^4.2.3",
+    "rollup-plugin-sourcemaps": "^0.4.2",
+    "rollup-plugin-terser": "^4.0.4",
+    "source-map-support": "^0.5.12",
+    "ts-node": "^8.1.0",
+    "typescript": "^3.4.5"
   },
-  "dependencies": {}
+  "dependencies": {
+    "strplace": "0.0.5"
+  }
 }
